old/requirements.txt
# Agency Performance Tracker - Dependencies

# Web Framework
streamlit>=1.28.0

# Database
sqlalchemy>=2.0.0

# Data manipulation
pandas>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Password hashing (authentication)
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0

# Charts (optional but recommended)
plotly>=5.18.0

# PostgreSQL adapter (for production - optional)
psycopg2-binary>=2.9.9

# File watching for development (optional)
# watchdog>=3.0.0
